+++ CHANGELOG.md
@@ -5,6 +5,10 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](http://keepachangelog.com/)
 and this project adheres to [Semantic Versioning](http://semver.org/).
 
+## [Unreleased]
+### Changed
+- Rename master branch to main [#67](https://github.com/IN-CORE/pyincore/issues/67)
+
 ## [1.2.0] - 2021-12-15
 
 ### Added