+++ CHANGELOG.md
@@ -6,8 +6,14 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/)
 and this project adheres to [Semantic Versioning](http://semver.org/).
 
 ## [Unreleased]
+
+### Added
+- Support for Restoration curves in pyincore. Included pytests. [#71](https://github.com/IN-CORE/pyincore/issues/71)
+- A DFR3 method to calculate the inverse of cdf which can be used in repair curve calculation to calculate repair time when percent of functionality is provided as an input. [#71](https://github.com/IN-CORE/pyincore/issues/71)
+
 ### Changed
 - Rename master branch to main [#67](https://github.com/IN-CORE/pyincore/issues/67)
+- Changed the DFR3 expression evaluation method to also accept an array of numeric values, instead of a single value. This improves analyses performance where an array of x values can be evaluated together. [#71](https://github.com/IN-CORE/pyincore/issues/71)
 
 ## [1.2.0] - 2021-12-15